+++ src/Client.php
@@ -25,7 +25,6 @@ use Berlioz\Http\Client\Exception\RequestException;
 use Berlioz\Http\Message\Request;
 use Berlioz\Http\Message\Response;
 use Berlioz\Http\Message\Uri;
-use Closure;
 use Psr\Http\Client\ClientExceptionInterface;
 use Psr\Http\Client\ClientInterface;
 use Psr\Http\Message\RequestInterface;
@@ -43,60 +42,30 @@ class Client implements ClientInterface, LoggerAwareInterface
     use Components\RequestFactoryTrait;
 
     private ?float $lastRequestTime = null;
+    private Options $options;
     private array $adapters;
     private Session $session;
 
     /**
      * Client constructor.
      *
-     * @param array $options
+     * @param Options|array|null $options
      * @param AdapterInterface ...$adapter
      *
      * @option string "baseUri" Base of URI if not given in requests
-     * @option bool "followLocation" Follow redirections (default: true)
-     * @option int "followLocationLimit" Limit location to follow
+     * @option false|int "followLocation" Follow redirections (default: 5)
      * @option int "sleepTime" Sleep time between requests (ms) (default: 0)
      * @option string "logFile" Log file name (only file name, not path)
      * @option bool "exceptions" Throw exceptions on error (default: true)
      * @option null|false|CookiesManager "cookies" NULL: to use default cookie manager; FALSE: to not use cookies; a CookieManager object to use
-     * @option callable "callback"  Callback after each request
-     * @option array  "headers" Default headers
+     * @option Closure "callback"  Callback after each request
+     * @option Closure "callbackException"  Callback on exception
+     * @option array "headers" Default headers
      */
-    public function __construct(protected array $options = [], AdapterInterface ...$adapter)
+    public function __construct(Options|array|null $options = null, AdapterInterface ...$adapter)
     {
-        // Merge with default options
-        $this->options = array_replace_recursive(
-            [
-                // Base of URI for URI requests
-                'baseUri' => null,
-                // If client follow locations
-                'followLocation' => true,
-                // Follow location limit in case of infinite loop
-                'followLocationLimit' => 5,
-                // Sleep time between each request
-                'sleepTime' => 0,
-                // File where log requests/responses
-                'logFile' => null,
-                // Throw exceptions on HTTP error cases
-                'exceptions' => true,
-                // NULL: to use default cookie manager; FALSE: to not use cookies; a CookieManager object to use
-                'cookies' => null,
-                // Callback on each request
-                'callback' => null,
-                // Default headers
-                'headers' => [
-                    'Accept' => ['text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8'],
-                    'User-Agent' => ['BerliozBot/1.0'],
-                    'Accept-Language' => ['fr,fr-fr;q=0.8,en-us;q=0.5,en;q=0.3'],
-                    'Accept-Encoding' => ['gzip, deflate'],
-                    'Accept-Charset' => ['ISO-8859-1,utf-8;q=0.7,*;q=0.7'],
-                    'Connection' => ['close'],
-                ],
-            ],
-            $this->options
-        );
-        $this->defaultHeaders = &$this->options['headers'];
-
+        $this->options = Options::make($options);
+        $this->defaultHeaders = &$this->options->headers;
         $this->adapters = $adapter ?: [extension_loaded('curl') ? new CurlAdapter() : new StreamAdapter()];
         $this->session = new Session();
     }
@@ -125,7 +94,17 @@ class Client implements ClientInterface, LoggerAwareInterface
         $this->adapters = $data['adapters'];
         $this->options = $data['options'];
         $this->session = $data['session'];
-        $this->defaultHeaders = &$this->options['headers'];
+        $this->defaultHeaders = &$this->options->headers;
+    }
+
+    /**
+     * Get options.
+     *
+     * @return Options
+     */
+    public function getOptions(): Options
+    {
+        return $this->options;
     }
 
     /**
@@ -186,24 +165,24 @@ class Client implements ClientInterface, LoggerAwareInterface
      *
      * @param RequestInterface $request
      * @param CookiesManager|false|null $cookies
-     * @param array $options
+     * @param Options $options
      *
      * @return RequestInterface
      */
     protected function prepareRequest(
         RequestInterface $request,
-        CookiesManager|false|null $cookies = null,
-        array $options = []
+        CookiesManager|false|null $cookies,
+        Options $options
     ): RequestInterface {
         $request = $request->withUri(
             Uri::create(
                 $request->getUri(),
-                $options['baseUri'] ?? $this->getSession()->getHistory()->getLast()?->getRequest()->getUri()
+                $options->baseUri ?? $this->getSession()->getHistory()->getLast()?->getRequest()->getUri()
             )
         );
 
         // Add default headers to request
-        $request = $this->addDefaultHeaders($request, $options['headers'] ?? []);
+        $request = $this->addDefaultHeaders($request, $options->headers);
 
         // Add content length
         if (false === $request->hasHeader('Content-Length')) {
@@ -226,40 +205,28 @@ class Client implements ClientInterface, LoggerAwareInterface
     /**
      * @inheritDoc
      */
-    public function sendRequest(RequestInterface $request, array $options = []): ResponseInterface
+    public function sendRequest(RequestInterface $request, Options|array $options = []): ResponseInterface
     {
         $followLocationCounter = 0;
         $originalRequest = $request;
 
         // Merge options with global options
-        $options = array_replace_recursive($this->options, $options);
+        $options = Options::make($options, $this->options);
 
         // Cookies manager
         // If option "cookies" is defined:
         // - false: no cookies sent in request
         // - CookieManager: no cookies sent in request
-        $cookies = $this->getSession()->getCookies();
-        if (isset($options['cookies'])) {
-            $cookies = $options['cookies'] ?: new CookiesManager();
-
-            if (!is_object($cookies) || !$cookies instanceof CookiesManager) {
-                throw new HttpClientException(
-                    sprintf(
-                        'Option "cookies" must be an instance of "%s" class, or null or false value',
-                        CookiesManager::class
-                    )
-                );
-            }
-        }
+        $cookies = $options->cookies ?? $this->getSession()->getCookies() ?: new CookiesManager();
 
         do {
-            $this->sleep();
+            $this->sleep($options);
             $request = $this->prepareRequest($request, $cookies, $options);
-            $adapter = $this->getAdapter($options['adapter'] ?? null);
+            $adapter = $this->getAdapter($options->adapter);
 
             try {
                 $this->lastRequestTime = microtime(true);
-                $response = $adapter->sendRequest($request);
+                $response = $adapter->sendRequest($request, $options->context);
 
                 // Add request to history
                 $this->getSession()->getHistory()->add($cookies, $request, $response, $adapter->getTimings());
@@ -274,9 +241,8 @@ class Client implements ClientInterface, LoggerAwareInterface
             $cookies->addCookiesFromResponse($request->getUri(), $response);
 
             // Callback
-            $callback = $options['callback'] ?? null;
-            if ($callback instanceof Closure) {
-                $callback($request, $response);
+            if (null !== $options->callback) {
+                ($options->callback)($request, $response);
             }
 
             $followLocation = false;
@@ -298,10 +264,10 @@ class Client implements ClientInterface, LoggerAwareInterface
             }
 
             // Follow location ?
-            if (false === $options['followLocation']) {
+            if (false === $options->followLocation) {
                 continue;
             }
-            $followLocation = ($followLocationCounter++ <= ($options['followLocationLimit'] ?? 5));
+            $followLocation = ($followLocationCounter++ <= $options->followLocation);
             if (!$followLocation) {
                 throw new RequestException('Too many redirection from host', $originalRequest);
             }
@@ -325,18 +291,24 @@ class Client implements ClientInterface, LoggerAwareInterface
             $request = $this->prepareRequest(
                 new Request(Request::HTTP_METHOD_GET, Uri::create($redirectUri, $request->getUri())),
                 $cookies,
-                array_replace($options, ['headers' => ['Referer' => (string)$request->getUri()]])
+                Options::make(['headers' => ['Referer' => (string)$request->getUri()]], $options),
             );
         } while ($followLocation);
 
         // Exceptions if error?
-        if ($options['exceptions']) {
+        if (true === $options->exceptions) {
             if (intval(substr((string)$response->getStatusCode(), 0, 1)) > 3) {
-                throw new HttpException(
+                $exception = new HttpException(
                     sprintf('%d - %s', $response->getStatusCode(), $response->getReasonPhrase()),
                     $originalRequest,
                     $response
                 );
+
+                if (null === $options->callbackException) {
+                    throw $exception;
+                }
+
+                return ($options->callbackException)($exception, $options);
             }
         }
 
@@ -346,7 +318,7 @@ class Client implements ClientInterface, LoggerAwareInterface
     /**
      * Sleep between two request.
      */
-    protected function sleep(): void
+    protected function sleep(Options $options): void
     {
         // Initial request
         if (null === $this->lastRequestTime) {
@@ -354,7 +326,7 @@ class Client implements ClientInterface, LoggerAwareInterface
         }
 
         // No sleep time
-        if (($sleepMicroTime = $this->options['sleepTime'] * 1000) <= 0) {
+        if (($sleepMicroTime = $options->sleepTime * 1000) <= 0) {
             return;
         }